requirements.txt
supabase==2.4.1
python-dotenv==1.0.1
langchain-openai==0.1.3
langchain==0.1.16